# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Roskisprojekti API
  version: 1.0.0
servers:
  - url: http://localhost:8080/api
paths:
  /sites:
    get:
      summary: Hae kaikki säiliöt
      responses:
        "200":
          description: Lista säiliöistä
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Site"

    post:
      summary: Luo uusi säiliö
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSiteRequest"
      responses:
        "200":
          description: Luotu säiliö
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Site"
  /sites/{id}:
    put:
      summary: Päivitä säiliö
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSiteRequest"
      responses:
        "200":
          description: Päivitetty säiliö
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Site"

    delete:
      summary: Poista säiliö
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Poisto onnistui
...
components:
  schemas:
    Site:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        location:
          type: string
        fillPercent:
          type: number
        capacity:
          type: number
    CreateSiteRequest:
      type: object
      required: [location]
      properties:
        name:
          type: string
        location:
          type: string
        capacity:
          type: number
    UpdateSiteRequest:
      type: object
      properties:
        name:
          type: string
        location:
          type: string
        capacity:
          type: number
...